requirements.txt
flask
flask-cors
flask-login
werkzeug
opencv-python
numpy
psycopg2-binary
pyotp
qrcode
Pillow
